% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-balance.R
\name{C1}
\alias{C1}
\title{Entropy of Class Proportions (C1)}
\usage{
C1(y)
}
\arguments{
\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The entropy of class proportions of the input data set.
}
\description{
Entropy of Class Proportions (C1)
}
