% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-balance.R
\name{C2}
\alias{C2}
\title{The Imbalance Ratio (C2) of a Multiclass Data Set.}
\usage{
C2(y)
}
\arguments{
\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The imbalance ratio of the input data set.
}
\description{
\code{C2} computes the imbalance ratio of a binary or multiclass data set.
This is not equivalent to IR, the imbalance ratio for binary data sets.
}
