% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-measures.R
\name{corr_abs}
\alias{corr_abs}
\title{Mean Absolute Correlation Coefficient.}
\usage{
corr_abs(x, y)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The mean absolute correlation coefficient of the input data set.
}
\description{
\code{corr_abs} computes the correlations between all pairs of features
for each class and then averages these values for all pairs and all classes.
}
