% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcme.R
\name{dcme}
\alias{dcme}
\title{Double-Complier Mediation Effect with Instrumental Variables}
\usage{
dcme(
  data,
  W,
  A,
  L,
  Z,
  M,
  Y,
  family = c("binomial", "gaussian"),
  folds = 1,
  g_learners = "SL.glm",
  q_learners = "SL.glm",
  p_learners = "SL.glm",
  c_learners = "SL.glm",
  mu_learners = "SL.glm"
)
}
\arguments{
\item{data}{A data.frame}

\item{W}{Name of baseline confounders}

\item{A}{Name of instrument for Z}

\item{L}{Name of instrument for M}

\item{Z}{Name of binary exposure}

\item{M}{Name of binary mediator}

\item{Y}{Name of outcome variable}

\item{family}{Variable type for outcome, Y. Options are "binomial" or "gaussian"}

\item{folds}{Number of folds for crossfitting}
}
\value{
A list of parameter estimates and confidence intervals.
}
\description{
Double-Complier Mediation Effect with Instrumental Variables
}
\examples{
TO DO
}
