% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap-of-individual-feature-values.R
\name{F1}
\alias{F1}
\title{Fisher's Discriminant Ratio (F1).}
\usage{
F1(x, y)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The Fisher's Discriminant Ratio of a binary data set.
}
\description{
\code{F1} computes the Fisher's Discriminant Ratio of a binary data set.
}
