% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap-of-individual-feature-values.R
\name{F2}
\alias{F2}
\title{Volume of Overlap Region (F2).}
\usage{
F2(x, y)
}
\arguments{
\item{x}{A data frame containing the features of the data.}

\item{y}{A vector or factor determining the class of each example in
\code{x}.}
}
\value{
The Volume of Overlap Region of a binary data set.
}
\description{
\code{F2} computes the Volume of Overlap Region of a binary data set.
}
