% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap-of-individual-feature-values.R
\name{F3}
\alias{F3}
\title{Maximum Individual Feature Efficiency (F3)}
\usage{
F3(x, y)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The maximum individual efficiency of each feature in separating the classes.
}
\description{
Maximum Individual Feature Efficiency (F3)
}
