% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{IR}
\alias{IR}
\title{The Imbalance Ratio (IR) of a Data Set.}
\usage{
IR(y)
}
\arguments{
\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The imbalance ratio of the input data set.
}
\description{
\code{IR} computes the imbalance ratio of a binary data set. It is defined
as the ratio of the number of majority class examples to the number of
minority class examples.
}
