% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability-of-classes.R
\name{N2}
\alias{N2}
\title{Ratio of Average Intra/Inter Class NN Distance.}
\usage{
N2(x, y)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The ratio between the average intra-class and average inter-class
 1-NN Euclidian distance.
}
\description{
\code{N2} computes the average intra-class and average inter-class distance
of the 1-NN, and returns the intra/inter ratio.
}
