% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability-of-classes.R
\name{N3}
\alias{N3}
\title{Error Rate of 1NN Classifier.}
\usage{
N3(x, y)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing (only) the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The leave-one-out error estimate of the 1NN classifier for the
 input data set.
}
\description{
\code{N3} computes the error rate of the 1NN classifier using the
leave-one-out strategy.
}
