% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-topology-and-density-of-manifolds.R
\name{N4}
\alias{N4}
\title{Nonlinearity of the 1-NN Classifier.}
\usage{
N4(x, y, n = 100)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}

\item{n}{Number of test points to generate from the input data set.}
}
\value{
The 1-NN error computed on a test set generated by linear
 interpolation of points in the input data set.
}
\description{
\code{N4} computes the 1-NN error on a test set generated by linear
interpolation of points of the same class.
}
