% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{num_classes}
\alias{num_classes}
\title{The Number of Classes in the Data Set.}
\usage{
num_classes(y)
}
\arguments{
\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The number of classes present in the input data set.
}
\description{
\code{num_classes} computes the number of classes present in the input
data set.
}
