% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{num_examples_minority}
\alias{num_examples_minority}
\title{The Number of Observations in the Minority Class.}
\usage{
num_examples_minority(x, y)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing (only) the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The number of examples in the minority class.
}
\description{
\code{num_examples_minority} computes the number of examples in the
minority class. It assumes the number of classes in the input data set is
two.
}
