% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{num_features}
\alias{num_features}
\title{The Number of Features in the Data Set.}
\usage{
num_features(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The number of features present in the input data set.
}
\description{
\code{num_features} computes the number of features present in the input
data set. It is just a wrapper for the \code{ncol} function.
}
