% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{num_features_binary}
\alias{num_features_binary}
\title{The Number of Binary Features in the Data Set.}
\usage{
num_features_binary(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The number of binary features present in the input data set.
}
\description{
\code{num_features_binary} computes the number of binary features present
in the input data set. NA values are ignored.
}
