% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{num_features_categorical}
\alias{num_features_categorical}
\title{The Number of Categorical Feautures in the Data Set.}
\usage{
num_features_categorical(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing (only) the features of the data.}
}
\value{
The number of categorical features present in the input data set.
}
\description{
\code{num_features_categorical} computes the number of categorical
features present in the input data set.
}
