% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{num_features_numeric}
\alias{num_features_numeric}
\title{The Number of Numeric Features in the Data Set.}
\usage{
num_features_numeric(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The number of numeric features present in the input data set.
}
\description{
\code{num_features_numeric} computes the number of numeric features
present in the input data set.
}
