% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{pca_dimension}
\alias{pca_dimension}
\title{PCA Dimension}
\usage{
pca_dimension(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The PCA dimension.
}
\description{
PCA dimension is the number of PCA components needed to represent 95%
of the data variance.
}
