% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{proportion_examples_majority}
\alias{proportion_examples_majority}
\title{The Proportion of Majority Examples in the Data Set.}
\usage{
proportion_examples_majority(y)
}
\arguments{
\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The proportion of examples that belong to the majority class.
}
\description{
\code{proportion_examples_majority} computes the proportion of examples
that belong to the majority class.
}
