% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{proportion_examples_minority}
\alias{proportion_examples_minority}
\title{The Proportion of Minority Examples in the Data Set.}
\usage{
proportion_examples_minority(y)
}
\arguments{
\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The proportion of examples that belong to the minority class.
}
\description{
\code{proportion_examples_minority} computes the proportion of examples
that belong to the minority class.
}
