% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-measures.R
\name{proportion_features_categorical}
\alias{proportion_features_categorical}
\title{The Proportion of Categorical Features in the Data Set.}
\usage{
proportion_features_categorical(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The proportion of features that are categorical in the input data
 set.
}
\description{
\code{proportion_features_categorical} computes the proportion of features
that are categorical.
}
