% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-measures.R
\name{sd_ratio}
\alias{sd_ratio}
\title{The Geometric Mean Ratio of Standard Deviations.}
\usage{
sd_ratio(x, y)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}

\item{y}{A vector (or factor) determining the class of each example in
the data.}
}
\value{
The geometric mean ratio of standard deviations.
}
\description{
\code{sd_ratio} computes the lack of homogeneity between the covariance
matrices of the classes in the data set.
}
