% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_x_and_y}
\alias{split_x_and_y}
\title{Split a Data Set into Predictors and Target.}
\usage{
split_x_and_y(data, target = ncol(data))
}
\arguments{
\item{data}{A data frame containing both predictors and target variables.}

\item{target}{The column index or the name of the target variable. If not
provided the target is assumed to be the last column of \code{data}.}
}
\value{
A list containing the predictors (x) and the target (y) variables.
}
\description{
\code{split_x_and_y} splits the input data set into the predictors x
and the target y.
}
