% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-topology-and-density-of-manifolds.R
\name{T2}
\alias{T2}
\title{Average Number of Points per Dimension (T2).}
\usage{
T2(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The average number of points per dimension of the input data set.
}
\description{
\code{T2} computes the average number of points per dimension of the input
data set.
}
