% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{T3}
\alias{T3}
\title{Average Number of Points per PCA Dimension (T3).}
\usage{
T3(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The average number of points per PCA dimension of the input data set.
}
\description{
\code{T3} computes the average number of points per PCA dimension of the input
data set. PCA dimension is the number of PCA components needed to represent 95%
of the data variance.
}
