% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{T4}
\alias{T4}
\title{Ratio of the PCA Dimension to the Original Dimension (T4).}
\usage{
T4(x)
}
\arguments{
\item{x}{A data frame, or an object that can be converted to a data frame,
containing only the features of the data.}
}
\value{
The ratio of the PCA dimension to the original dimension.
}
\description{
\code{T4} computes the proportion of relevant dimensions of the dataset.
}
